Add openstack-helm repo images
This adds the building of the images present in the OpenStack-Helm
repository.

Images follow the same process as images from the osh-infra
repository.

Kubernetes-entrypoint is not used anymore and therefore was
not migrated over.

Likewise, LOCI is not following these images building process,
and therefore will be included in a different patch.

The manual build script do not expose the extra arguments
previously exposed, like KUBE_VERSION or OVS_VERSION by default.
It also means a future tagged version of this repository will
use the KUBE_VERSION or OVS_VERSION that are in tree, and
therefore very explicit.

Change-Id: If4ff73c37ed8bccb9f6951f3ccdbb7640e9b55a5