Publish images on DockerHub
This adds a playbook for publish jobs which will find
"openstackhelm" images, tag them for dockerhub publication,
and upload them based on booleans.

These publish jobs are put in a post merge pipeline.
Post merge pipeline should still have the content of the files changed
in the patches, and therefore can be still conditionally triggered
with the files: directive.

Change-Id: I134e34815743a87b003e1d1fb78974937ae2392e