Fix periodics
Periodics don't get scheduled if 'files' directive is used in
the job definition.

This is a problem, as we wanted to use files to restrict the
amount of images to use while gating.

The solution is to have two kind of jobs: check/gates jobs with a
more surgical approach, and periodic/post jobs with a global
approach.

This patch creates unnamed "variants" for check/gates that are
based on the periodic/post jobs, adding only the 'file' directive.

This patch also sorts the jobs in zuul files alphabetically, so
it is easier to find them.

The projects's pipelines are also sorted according to a
chronological order: check/gates/post/periodics jobs.

LOCI jobs were also fixed in this commit.
Some jobs were passing but were inconsistently configured
(voting in checks but not running in gates). They are now
moved to voting/gating.

Change-Id: Ifa28d894167003affc71256e49805b3482bf9a61