libvirt: Stick to 4.0.0 version located in main repository
Version 5 of libvirt present in Virtualization repo is too recent
and unnecessary for queens and rocky builds.

Virtualization repo is however necessary for cgroups packages
that are not present in the main repository. It can be deleted once the
necessary packages are installed.

Ceph repo should not be needed at all.

Change-Id: I2d030c0f326cef4571ef6d15056d8afdf7381d69