mini-mirror: Allow per-source Aptly config files
Current mini-mirror images are build using a single Aptly config file
provided at build time. In some cases, it may be desirable to provide an
Aptly configuration file for each source to change the behavior of a
single mirror. This commit introduces support for providing an Aptly
configuration file for each individual source. The config file should be
named aptly.conf and placed in the root source path.

Change-Id: Ic51fa2ab6898eaa14b45b67ba6708f2d7aea75eb