Publish LOCI images
Currently the OpenStack images are not published on dockerhub in
openstackhelm namespace.

This makes use of the standard openstack infra roles as much as
we can, to build and promote images.

The only thing that differs is the creation of the requirements
image, which needs to be available before building
the other images. For this, we:
1) Build the requirements container using infra role.
2) Build and run our own docker webserver container, extracting
   the wheels from the requirements container.

Depends-On: https://review.openstack.org/640711/

Change-Id: I16fc6e989f1c412d458270195c6aff8343524c48