mini-mirror: Ignore src config for publishing
Currently, mini-mirror honors per-source configuration files for the
creation of mirrors and publishing; however, some options exposed by
Aptly have a different impact depending on the operation. This commit
removes the per-source configuration file option from the publishing
process and defaults to the aptly config file supplied for all sources
in order to circumvent this issue.

Change-Id: I87659725918e776482aabccb97195216047f4c77