Do not hardcode libvirt version
This means any security update of libvirt would break the image
(1.3.1-1ubuntu10.26 was released recently, and broke the build).

Instead, we should rely on the fact packagers are doing the right
thing with their version.

Should we want to freeze to certain versions, then we should
provide a ppa and deal with the appropriate packaging of libvirt.

Change-Id: Ia7bba83ae5610a91267bae4d8b80c4c1b8dab70c