Remove Newton from image builds
During the last PTG, we discussed about removing the support of
Newton image building. As nobody was interested in maintaining
newton, there is no reason to test this anymore.

Without this patch, we'll continue to CI test Newton, and should
a failure happen, nobody will care to fix it. It has therefore
no purpose.

To not break compatibility or existing tooling, this patch
removes Newton from CI builds, and doesn't remove the wrapper
script we provide in openstack/loci/.

Change-Id: If2d60001da03930064afa4d873d8597aeb8f2ee7