Restrict Rocky on Leap 15.0
The default base image for opensuse/leap:15 is now pointing
to 15.1 on dockerhub.

In LOCI, we are using opensuse/leap:15 as base image by
default. It is not a problem, as we are always targetting
master in LOCI.

However, Rocky is older, and the upper-constraints for
libvirt currently points to 4.6.0, which is unable to build
on leap 15.1, due to 15.1 having a too recent version of
libvirt.

This fixes the requirements building by ensuring Rocky
will always stay on a 15.0 base.

Change-Id: I033f1f5dc39c47b748b3887f25e1dfa02fc0f800