Add jq package for neutron image
In neutron-openvswitch-init.sh script, auto_bridge_add json file is
parsed. As of now the json file is very simple, we parse it with command
'sed' and 'tr'[1]. While we are working on ovs-dpdk support, which makes
the auto_bridge_add json file more complicated(pci_id, n_rxq, etc.).

This patch is to include jq package to parse json.

[1]
https://github.com/openstack/openstack-helm/blob/dadf9946e076df2b09556f4a18107dc487788cdd/neutron/templates/bin/_neutron-openvswitch-agent-init.sh.tpl#L40

Change-Id: I678cdaf4a7f2072d5dd9bb8618e0beb8f3a05655