Freeze LOCI images with a date
In the last patch to freeze images with a date, the LOCI images
were not added, do to the length of the patch [1].

This patch fixes the absence of LOCI based frozen image by
changing the CI appropriately, in the same way as the previous
patch ([1]).

However, LOCI configuration here was most of the time relying
on inheritance, which made bugs hard to find. This cleans up
the LOCI building by:
1) Having a common abstract profile for which all the main
   zuul variables are inherited
2) Making sure each branch has content without inheritance,
   but with only limited repetitions. This uses anchors
   to reduce the information to type about _projects, as
   the project details are most likely the same as master.

[1]: https://review.opendev.org/#/c/663744

Depends-On: https://review.opendev.org/676398
Change-Id: I6f33d5b7a33904cdd1c5ec352046ca9a9bca2aba