mini-mirror: Use latest aptly version
Currently, the mini-mirror image job installs aptly from the standard
Ubuntu repositories. Consequently, the version of aptly used for the
Xenial build is 0.9.1 and differs from the version of aptly in the
Bionic build.

This change adds the official aptly repositories when installing aptly
in order to use the latest version and maintain consistency between the
two builds.

Change-Id: Ic2f79b0a910f0c0e3e2aff361dfe14c9b312f0f9
Signed-off-by: Drew Walters <andrew.walters@att.com>