Nagios; Update check_exporter_health_metric plugin
This updates the Nagios plugin for checking health metrics exposed
by an exporter endpoint directly. To address issues where multiple
exporter replicas may not all be active, this moves the plugin to
instead use the python kubernetes client to programmatically
determine which endpoint tied to the exporter service is active
and returns the metrics exposed by the active endpoint. This
allows for more robust checking of service health in scenarios
where circumventing prometheus is desired

Change-Id: I14e21936d1808a4f41b20368451da95100075dda
Signed-off-by: Steve Wilkerson <sw5822@att.com>