mini-mirror: Preserve codename and label
Ubuntu 18.04 performs strict checking of the Codename and Label when
performing an apt-get update. The snapshots published by aptly do not
set these parameters in a way that apt-get likes.

This change explicitly sets the codename and label based on the values
from the upstream repo being mirrored.

Note: this relies on a currently unmerged pull request/
https://github.com/aptly-dev/aptly/pull/892

Change-Id: I011862dcf9efbb8b1f9757c65f3d58da49a496d7