Fix source publishing and publish logic
Sometime recently the logic in aptly changed so that even if the
configuration option "downloadSourcePackages" is true, the source
packages will not be published in the resulting mirror unless the
architectures list includes "source", but adding "source" to the
actual architecture list for a repo will prevent the mirror creation
from functioning as most repos don't have a "source" architecture.

This change adds a new check to look at the aptly configuration file to
determine if "downloadSourcePackages" is true, and if so then it will
add a flag to the publish command with the architectures from the
configuration plus the "source" architecture.

I also modified the publish command to bring it into one command while
still allowing it to be configurable as it was before.

This also removes the "set -x" to limit the excessive output

Change-Id: Iebf1e862cb0a48a5a5eeda6ea0e8df3537eb602b