Append the domain name to the host name in Nagios dashboard
This adds logic to check if an environment variable has been
set for the domain name and appends it to the host name
if it has so that the full FQDN appears in Nagios dashboard.
If the the domain name has not been set just the host name is
given like previous behavior.

Change-Id: Id42edb073d4701ddb61f4957af7e5ac5f931dfbf