[Mariadb] Add pykube-ng to maria
Right now we are using nginx ingress as LB on top of mariadb, this adds
aditional ovehead between client and server. Additionally there is no
way to control master node with ingress, which creates backup procedure
more complex. To avoid that alternative approach is proposed
https://review.opendev.org/#/c/752679/, which allows to drop ingress.



pykube-ng is needed to allow connecting to kubernetes from
mariadb-controller pod that selects master node.

Change-Id: I4e227c8c20e7409d1fc77d71ee0159aff5e07979