Minimize ceph-daemon image size
These changes are to minimize the ceph-daemon image size as is
recommended by industry standard best practices. The main changes
are the removal of various debugging tools such as:

valgrind
hexedit
sysstat
dstat
iperf

as well as compilers and debuggers

gcc
gcc-7
cpp-7
gdb

Etcdctl and forego are also removed.

Change-Id: I9502c5f297b1bfcdaafb82ae438f3ba6382c2c64