Align mini-mirror Dockerfile args with build.sh
The mini-mirror build.sh script has somehow ended up out of sync with
the Dockerfiles, and is not passing the correct build-args to allow base
image overrides.

This change aligns with the standard of using BUILD_FROM (ubuntu) for
the build stage and FROM (nginx) for the final stage.

Relevant earlier changes:
0: https://review.opendev.org/714002
1: https://review.opendev.org/679315

Change-Id: I844efa669c0f11004a44f20f6725338e1d2699b6