[fix] Openstack exporter update
This ps updates the logic to clear the token when it
receives authorization error from keystone.
This is a special case of 401.
When the exporter receives a 401 error from the keystone and the
application is not placement, we reset the token value such that on the
next retry it reinitialize the parameters.

Change-Id: I0c4bcb17aa420e31591405b6000a2db2821e2428