Update mariadb upload job
A previous change[0] added an ubuntu bionic Dockerfile for mariadb
but did not update the zuul file to properly tag the new image.

This change updates the zuul job in order to correctly tag the image,
as well as simplifying the ability to specify the ubuntu release.

Also updates the previous change's bionic Dockerfile to focal.

[0] https://review.opendev.org/c/openstack/openstack-helm-images/+/782843

Change-Id: I0c803bf1d2d4705b669a48bba0b9576c71d4b77d