Add Focal images for Libvirt
This patch adds images using Focal for Libvirt, primarily, the
libvirt-bin package which was a transitionary package is now
installing the explicit daemon and client packages and the ipxe-qemu
includes all the ROMs for iPXE booting.

It also drops the Xenial images as they are no longer supported
and quite old now.

Change-Id: Ic0a2ad8610c383dbbe8b71020612d11dc2f1040e