Comment out train release jobs
The recent addition of the vpn profile seems to break train with
an issue in neutron-vpnaas that tries to find the ussuri release
in train.

Since train is in extendend maintenance anyway and Yoga about to
come out, this change comments out the train jobs to unblock the
gate. Further discussions will happen about potentially dropping
support for Train.

Change-Id: I3242c9244e139fffccde031df1de637e9929e525