Remove heat-dashboard from default image building
Currently heat-dashboard is failing to build with dependency issues
on the Yoga release and the repo has been taking quite some time
to resolve this.

This change removes the inclusion of heat-dashboard from the default
image building for horizon and requirements, opting to include it
in the specific releases instead.

Change-Id: I49e549d78334cd9079e263766d15672a5d455408