Remove minikube-aio image building/jobs
With the move to helm v3, we have since stopped using the
minikube-aio image and playbooks in favor of a single script
to deploy kubernetes in the zuul checks and gates. This change
removes the image building and jobs for minikube-aio and removes
any dependencies that other jobs had on it.

Change-Id: Icd6f801eb43778bf75a16cb1e83f6534423a58db