Disable X & Y release jobs temporarily
Currently the compute-kit jobs for openstack xena and yoga have issues
running within Zuul and timeout when trying to exercise nova and neutron.

This change comments out the X & Y release jobs while we diagnose and fix
the issue. A future change will uncomment the jobs so they run again.

Change-Id: Ic826c0ebf9b0ad82598b729c2c16eaf93769e3d2