Fix image promote job definition
If secrets are defined in a job they
are only available to the playbooks
defined in the same job. They are not available
to any child jobs and they only available to
the parent jobs if pass-to-parent is true.

In our case we use custom promote playbook thus
we have to define secrets in the same job.

Change-Id: I5c60a1d109ad8cb55f3d7f61c74653d4590e3942