Fix libvirt Dockerfile
By default all Openstack deployment test jobs
use latest-ubuntu_focal image which has libvirt
version 6.0.0 and qemu version 4.2 which is
incompatible with Bobcat. This old image was published
more than a year ago and since that has never been
updated and it also contains the nova user.

The Openstack deployment with newer libvirt
images fails due to this
https://review.opendev.org/c/openstack/openstack-helm-images/+/864772

This PR adds nova user which is necessary for libvirt
to successfully communicate with Nova.

Change-Id: I0d5f302261611b0666323178a52ca974126be1a5